requirements.txt
flask
python-dotenv
line-bot-sdk
openai
requests